Support Python 3
