+++ mcache/tests/conftest.py
@@ -1,93 +1,60 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-
-import subprocess
 import os
-import tempfile
-import shutil
-import time
-import pytest
+
 import bmemcached
-from bmemcached.exceptions import MemcachedException
+import pytest
 
+from datadog_checks.dev import TempDir, WaitFor, docker_run
 from datadog_checks.mcache import Memcache
-
-from common import (HERE, PORT, HOST, USERNAME, PASSWORD, DOCKER_SOCKET_DIR, DOCKER_SOCKET_PATH)
-
-
-@pytest.fixture(scope="session")
-def memcached():
-    """
-    Start a standalone Memcached server.
-    """
-    env = os.environ
-    env['PWD'] = HERE
-    docker_compose_file = os.path.join(HERE, 'compose', 'docker-compose.yaml')
-
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "up", "-d", "memcached"], env=env)
-    attempts = 0
-    while True:
-        if attempts > 10:
-            raise Exception("Memcached boot timed out!")
-
-        mc = bmemcached.Client(["{}:{}".format(HOST, PORT)], USERNAME, PASSWORD)
-        try:
-            mc.set("foo", "bar")
-        except MemcachedException:
-            attempts += 1
-            time.sleep(1)
-        else:
-            mc.delete("foo")
-            mc.disconnect_all()
-            break
-
-    yield
-
-    subprocess.check_call(["docker-compose", "-f", docker_compose_file, "down"])
+from .common import (
+    HERE, PORT, HOST, USERNAME, PASSWORD, DOCKER_SOCKET_DIR, DOCKER_SOCKET_PATH, platform_supports_sockets
+)
+from .utils import get_host_socket_path
 
 
-@pytest.fixture(scope="session")
-def memcached_socket():
-    """
-    Start a standalone Memcached server.
-    """
+def connect_to_mcache(*client_args):
+    client = bmemcached.Client(*client_args)
+
     try:
-        host_socket_dir = os.path.realpath(tempfile.mkdtemp())
-        host_socket_path = os.path.join(host_socket_dir, 'memcached.sock')
-        os.chmod(host_socket_dir, 0o777)
-
-        env = os.environ
-        env['PWD'] = HERE
-        env['DOCKER_SOCKET_DIR'] = DOCKER_SOCKET_DIR
-        env['DOCKER_SOCKET_PATH'] = DOCKER_SOCKET_PATH
-        env['HOST_SOCKET_DIR'] = host_socket_dir
-
-        docker_compose_file = os.path.join(HERE, 'compose', 'docker-compose.yaml')
-        subprocess.check_call(["docker-compose", "-f", docker_compose_file, "down"])
-        subprocess.check_call(["docker-compose", "-f", docker_compose_file, "up", "-d", "memcached_socket"], env=env)
-
-        attempts = 0
-        while True:
-            if attempts > 10:
-                raise Exception("Memcached boot timed out!")
-
-            mc = bmemcached.Client(host_socket_path, USERNAME, PASSWORD)
-            try:
-                mc.set("foo", "bar")
-            except MemcachedException:
-                attempts += 1
-                time.sleep(1)
-            else:
-                mc.delete("foo")
-                mc.disconnect_all()
-                break
-
-        yield host_socket_dir, host_socket_path
-
-        subprocess.check_call(["docker-compose", "-f", docker_compose_file, "down"])
+        client.set('foo', 'bar')
+        client.delete('foo')
     finally:
-        shutil.rmtree(host_socket_dir, ignore_errors=True)
+        client.disconnect_all()
+
+
+@pytest.fixture(scope='session')
+def dd_environment(e2e_instance):
+    with docker_run(
+        os.path.join(HERE, 'compose', 'docker-compose.yaml'),
+        service_name='memcached',
+        env_vars={'PWD': HERE},
+        conditions=[WaitFor(connect_to_mcache, args=(['{}:{}'.format(HOST, PORT)], USERNAME, PASSWORD))]
+    ):
+        if platform_supports_sockets:
+            with TempDir() as temp_dir:
+                host_socket_path = os.path.join(temp_dir, 'memcached.sock')
+
+                if not os.path.exists(host_socket_path):
+                    os.chmod(temp_dir, 0o777)
+
+                with docker_run(
+                    os.path.join(HERE, 'compose', 'docker-compose.yaml'),
+                    service_name='memcached_socket',
+                    env_vars={
+                        'DOCKER_SOCKET_DIR': DOCKER_SOCKET_DIR,
+                        'DOCKER_SOCKET_PATH': DOCKER_SOCKET_PATH,
+                        'HOST_SOCKET_DIR': temp_dir,
+                        'HOST_SOCKET_PATH': host_socket_path,
+                    },
+                    conditions=[WaitFor(connect_to_mcache, args=(host_socket_path, USERNAME, PASSWORD))],
+                    # Don't worry about spinning down since the outermost runner will already do that
+                    down=lambda: None
+                ):
+                    yield e2e_instance
+        else:
+            yield e2e_instance
 
 
 @pytest.fixture
@@ -96,8 +63,8 @@ def client():
 
 
 @pytest.fixture
-def client_socket(memcached_socket):
-    return bmemcached.Client(memcached_socket[1], USERNAME, PASSWORD)
+def client_socket():
+    return bmemcached.Client(get_host_socket_path(), USERNAME, PASSWORD)
 
 
 @pytest.fixture
@@ -108,26 +75,30 @@ def check():
 @pytest.fixture
 def instance():
     return {
-        'url': "{}".format(HOST),
+        'url': HOST,
         'port': PORT,
-        'tags': ["foo:bar"],
+        'tags': ['foo:bar'],
         'username': USERNAME,
         'password': PASSWORD,
     }
 
 
-@pytest.fixture
-def instance_socket(memcached_socket):
+@pytest.fixture(scope='session')
+def e2e_instance():
     return {
-        'socket': memcached_socket[1],
-        'tags': ["foo:bar"],
+        'url': HOST,
+        'port': PORT,
+        'tags': ['foo:bar'],
         'username': USERNAME,
         'password': PASSWORD,
     }
 
 
 @pytest.fixture
-def aggregator():
-    from datadog_checks.stubs import aggregator
-    aggregator.reset()
-    return aggregator
+def instance_socket():
+    return {
+        'socket': get_host_socket_path(),
+        'tags': ['foo:bar'],
+        'username': USERNAME,
+        'password': PASSWORD,
+    }
