+++ mcache/tests/test_integration.py
@@ -1,44 +1,16 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-
 import pytest
-import subprocess
-import os
 
-from datadog_checks.utils.platform import Platform
-
-from common import (PORT, SERVICE_CHECK, HOST, GAUGES, RATES, ITEMS_RATES, ITEMS_GAUGES, SLABS_RATES, SLABS_GAUGES,
-                    SLABS_AGGREGATES)
-
-
-def count_connections(port):
-    """
-    Count how many connections to memecached in the current process
-    """
-    pid = os.getpid()
-    p1 = subprocess.Popen(['lsof', '-a', '-p%s' % pid, '-i4'], stdout=subprocess.PIPE)
-    p2 = subprocess.Popen(["grep", ":%s" % port], stdin=p1.stdout, stdout=subprocess.PIPE)
-    p3 = subprocess.Popen(["wc", "-l"], stdin=p2.stdout, stdout=subprocess.PIPE)
-    output = p3.communicate()[0]
-    return int(output.strip())
+from .common import HOST, PORT, SERVICE_CHECK, requires_socket_support, requires_unix_utils
+from .metrics import GAUGES, RATES, ITEMS_GAUGES, ITEMS_RATES, SLABS_GAUGES, SLABS_RATES, SLABS_AGGREGATES
+from .utils import count_connections, get_host_socket_path
 
 
 @pytest.mark.integration
-def test_connections_leaks(check, instance, memcached):
-    """
-    This test was ported from the old test suite but the leak might not be a
-    problem anymore.
-    """
-    # Start state, connections should be 0
-    assert count_connections(PORT) == 0
-    check.check(instance)
-    # Verify that the count is still 0
-    assert count_connections(PORT) == 0
-
-
-@pytest.mark.integration
-def test_service_ok(check, instance, aggregator, memcached):
+@pytest.mark.usefixtures('dd_environment')
+def test_service_ok(check, instance, aggregator):
     """
     Service is up
     """
@@ -51,7 +23,8 @@ def test_service_ok(check, instance, aggregator, memcached):
 
 
 @pytest.mark.integration
-def test_metrics(client, check, instance, aggregator, memcached):
+@pytest.mark.usefixtures('dd_environment')
+def test_metrics(client, check, instance, aggregator):
     """
     Test all the available metrics: default, options and slabs
     """
@@ -79,16 +52,18 @@ def test_metrics(client, check, instance, aggregator, memcached):
     assert aggregator.metrics_asserted_pct == 100.0
 
 
-# This test doesn't work with Mac either.
-# See https://docs.docker.com/docker-for-mac/osxfs/#file-types
-# See open PR https://github.com/docker/for-mac/issues/483
+@requires_socket_support
 @pytest.mark.integration
-@pytest.mark.skipif(not Platform.is_linux(), reason='Windows sockets are not file handles')
-def test_service_with_socket_ok(check, instance_socket, aggregator, memcached_socket):
+@pytest.mark.usefixtures('dd_environment')
+def test_service_with_socket_ok(check, instance_socket, aggregator):
     """
     Service is up
+
+    This test doesn't work with Mac either.
+    See https://docs.docker.com/docker-for-mac/osxfs/#file-types
+    See open PR https://github.com/docker/for-mac/issues/483
     """
-    tags = ["host:unix", "port:{}".format(memcached_socket[1]), "foo:bar"]
+    tags = ["host:unix", "port:{}".format(get_host_socket_path()), "foo:bar"]
     check.check(instance_socket)
     assert len(aggregator.service_checks(SERVICE_CHECK)) == 1
     sc = aggregator.service_checks(SERVICE_CHECK)[0]
@@ -96,9 +71,10 @@ def test_service_with_socket_ok(check, instance_socket, aggregator, memcached_so
     assert sc.tags == tags
 
 
+@requires_socket_support
 @pytest.mark.integration
-@pytest.mark.skipif(not Platform.is_linux(), reason='Windows sockets are not file handles')
-def test_metrics_with_socket(client_socket, check, instance_socket, aggregator, memcached_socket):
+@pytest.mark.usefixtures('dd_environment')
+def test_metrics_with_socket(client_socket, check, instance_socket, aggregator):
     """
     Test all the available metrics: default, options and slabs
     """
@@ -115,7 +91,7 @@ def test_metrics_with_socket(client_socket, check, instance_socket, aggregator, 
     })
     check.check(instance_socket)
 
-    expected_tags = ["url:unix:{}".format(memcached_socket[1]), 'foo:bar']
+    expected_tags = ["url:unix:{}".format(get_host_socket_path()), 'foo:bar']
     for m in GAUGES + RATES + SLABS_AGGREGATES:
         aggregator.assert_metric(m, tags=expected_tags, count=1)
 
@@ -124,3 +100,18 @@ def test_metrics_with_socket(client_socket, check, instance_socket, aggregator, 
         aggregator.assert_metric(m, tags=expected_tags, count=1)
 
     assert aggregator.metrics_asserted_pct == 100.0
+
+
+@requires_unix_utils
+@pytest.mark.integration
+@pytest.mark.usefixtures('dd_environment')
+def test_connections_leaks(check, instance):
+    """
+    This test was ported from the old test suite but the leak might not be a
+    problem anymore.
+    """
+    # Start state, connections should be 0
+    assert count_connections(PORT) == 0
+    check.check(instance)
+    # Verify that the count is still 0
+    assert count_connections(PORT) == 0
