+++ mcache/tox.ini
@@ -2,23 +2,13 @@
 minversion = 2.0
 basepython = py27
 envlist =
+    py{27,36}-integration
     unit
-    integration
     flake8
 
 [testenv]
 usedevelop = true
 platform = linux|darwin|win32
-
-[testenv:unit]
-deps =
-    -e../datadog_checks_base[deps]
-    -rrequirements-dev.txt
-commands =
-    pip install --require-hashes -r requirements.txt
-    pytest -m"not integration" -v
-
-[testenv:integration]
 passenv =
     DOCKER*
     COMPOSE*
@@ -29,11 +19,16 @@ commands =
     pip install --require-hashes -r requirements.txt
     pytest -m"integration" -v
 
+[testenv:unit]
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest -m"not integration" -v
+
 [testenv:flake8]
 skip_install = true
 deps = flake8
 commands = flake8 .
 
 [flake8]
-exclude = .eggs,.tox
+exclude = .eggs,.tox,build
 max-line-length = 120
